requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
bigquery-magics==0.10.0
Bottleneck==1.4.2
branca==0.6.0
Brotli==1.0.9
cachetools==5.5.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.11.0
db-dtypes==1.4.3
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.13
distro==1.9.0
elastic-transport==8.17.0
elasticsearch==8.17.0
exceptiongroup==1.3.0
executing==2.2.0
fastavro==1.11.1
folium==0.20.0
fonttools==4.55.3
frozenlist==1.5.0
geopandas==1.0.1
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.34.0
google-cloud-bigquery-storage==2.31.0
google-cloud-core==2.4.3
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httpx==0.28.1
idna==3.7
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
mapclassify==2.9.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
multidict==6.1.0
nest_asyncio==1.6.0
networkx==3.4.2
numexpr==2.10.1
numpy==2.3.1
oauthlib==3.3.1
openai==1.77.0
opentelemetry-api==1.34.1
opentelemetry-instrumentation==0.48b0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.2.3
pandas-gbq==0.28.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.2
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pyzmq==27.0.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.7.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==72.1.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.0
stack_data==0.6.3
threadpoolctl==3.5.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
unicodedata2==15.1.0
urllib3==1.26.19
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xyzservices==2022.9.0
yarl==1.18.0
zipp==3.23.0
